+++ ICE.SDKtoAPI/LenderApiContractsV1/RateAdjustment.cs
@@ -9,6 +9,6 @@ namespace ICE.SDKtoAPI.LenderApiContractsV1
         public string AdjustmentType { get; set; }
         public string Description { get; set; }
         public string PriceAdjustmentType { get; set; }
-        public double? Adjustment { get; set; }
+        public decimal? Adjustment { get; set; }
     }
 }
